requirements.txt
absl-py==2.2.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aqtp==0.8.3
array-record==0.7.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
catkin-pkg==1.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.89
click==8.1.8
clip==1.0
cloudpickle==3.1.1
clu==0.0.12
coloredlogs==15.0.1
comm==0.2.2
configargparse==1.7
contourpy==1.3.1
cycler==0.12.1
cython==3.0.12
dash==3.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
distrax==0.1.5
distro==1.9.0
dm-tree==0.1.9
docstring-parser==0.16
docutils==0.21.2
editdistance==0.8.1
einops==0.8.1
equinox==0.11.12
etils==1.12.2
fastjsonschema==2.21.1
filelock==3.18.0
filterpy==1.4.5
flask==3.0.3
flatbuffers==25.2.10
flax==0.10.4
flaxformer==0.8.8
fonttools==4.56.0
freetype-py==2.5.1
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
h5py==3.13.0
humanfriendly==10.0
humanize==4.12.1
idna==3.10
imageio==2.37.0
immutabledict==4.2.1
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipython==9.1.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jax==0.5.3
jax-cuda12-pjrt==0.5.3
jax-cuda12-plugin==0.5.3
jaxlib==0.5.3
jaxopt==0.8.3
jaxtyping==0.3.0
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.14
keras==3.9.0
kiwisolver==1.4.8
kornia==0.8.0
kornia-rs==0.1.8
lazy-loader==0.4
libclang==18.1.1
lineax==0.0.7
lvis==0.5.3
mako==1.3.10
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-collections==1.0.0
ml-dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
namex==0.0.8
narwhals==1.35.0
nbformat==5.10.4
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.8.93
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnxruntime==1.21.1
open3d==0.19.0
opencv-python==4.8.0.74
openexr==3.3.5
opt-einsum==3.4.0
optax==0.2.5.dev0
optree==0.14.1
orbax-checkpoint==0.11.10
ott-jax==0.2.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
panopticapi==0.1
pillow==11.1.0
pip==25.1.1
platformdirs==4.3.7
plotly==6.0.1
pooch==1.8.2
promise==2.3
propcache==0.3.1
protobuf==3.20.1
pupil-apriltags==1.0.4.post11
pyarrow==19.0.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycryptodomex==3.23.0
pycuda==2025.1
pydeprecate==0.3.2
pyglet==2.1.6
pygments==2.19.1
pyopengl==3.1.0
pyopengl-accelerate==3.1.9
pyparsing==3.2.2
pypng==0.20220715.0
pyqt5==5.15.11
pyqt5-qt5==5.15.17
pyqt5-sip==12.17.0
pyqtgraph==0.13.7
pyquaternion==0.9.9
pyrender==0.1.45
python-gnupg==0.5.4
pytools==2025.1.5
pytorch-lightning==1.6.5.post0
pytz==2025.2
pyvista==0.45.2
pyvistaqt==0.11.2
pyyaml==6.0.2
qtpy==2.4.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rospkg==1.6.0
rpds-py==0.24.0
scenic==0.0.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
scooby==0.10.1
segment-anything==1.0
sentencepiece==0.2.0
simple-parsing==0.1.7
simplejson==3.20.1
siphash24==1.7
spicy==0.16.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.1
tensorflow-cpu==2.18.1
tensorflow-datasets==4.9.8
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.25.0
tensorflow-text==2.18.1
tensorstore==0.1.72
termcolor==2.5.0
tf-keras==2.18.0
tfds-nightly==4.9.8.dev202503240044
threadpoolctl==3.6.0
tifffile==2025.3.13
toml==0.10.2
toolz==1.0.0
torch==2.6.0
torchmetrics==0.6.0
torchvision==0.21.0
tqdm==4.67.1
treescape==0.1.9
trimesh==4.6.10
triton==3.2.0
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
vtk==9.4.2
wadler-lindig==0.1.4
werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.17.2
yacs==0.1.8
yarl==1.20.0
zipp==3.21.0